Change convert_puml_files.sh to use SHA1 checksums

+++ container_build/docker_puml/convert_puml_files.sh
@@ -47,6 +47,8 @@ debug() {
 convert_file() {
   local puml_file="$1"; shift
 
+  echo -e "\n${GREEN}Processing source file ${BLUE}${puml_file}${GREEN}"
+
   # TODO AT: It may be worth generating a .puml.sha1 for each .puml file.
   #   Then we can see if the .puml has changed and only regen it if there is
   #   no .puml.svg or the sha1 is different. This would speed up the process
@@ -58,32 +60,62 @@ convert_file() {
   # Replace first match starting at end
   local generated_svg_filename="${puml_filename/%\.puml/.svg}"
   local renamed_svg_filename="${puml_filename}.svg"
+  local sha1_filename="${puml_filename}.sha1"
 
   local puml_file_dir
   puml_file_dir="$(dirname "${puml_file}")"
 
   local generated_svg_file="${puml_file_dir}/${generated_svg_filename}"
   local renamed_svg_file="${puml_file_dir}/${renamed_svg_filename}"
+  local sha1_file="${puml_file_dir}/${sha1_filename}"
 
-  echo -e "${GREEN}Converting file ${BLUE}${puml_file}${GREEN}" \
-    "to ${BLUE}${renamed_svg_filename}${NC}"
+  local is_conversion_needed=false
 
-  local is_success=true
-  # convert the .puml to .svg with same name
-  java \
+  if [[ -f "${renamed_svg_file}" ]]; then
+    if [[ -f "${sha1_file}" ]]; then
+      # Busybox version of sha1sum so -s instead of --quiet
+      if sha1sum -c -s "${sha1_file}" >/dev/null 2>&1; then
+        echo -e "${GREEN}PUML file has not changed since last conversion${NC}"
+      else
+        echo -e "${GREEN}PUML file has changed, conversion required${NC}"
+        is_conversion_needed=true
+      fi
+    else
+      # No sha1 file
+      is_conversion_needed=true
+    fi
+  else
+    # No SVG file
+    is_conversion_needed=true
+  fi
+
+  if [[ "${is_conversion_needed}" = true ]]; then
+
+    echo -e "${GREEN}Converting PUML to ${BLUE}${renamed_svg_filename}${NC}"
+
+    local is_success=true
+    # convert the .puml to .svg with same name
+    java \
       -jar /builder/plantuml.jar \
       "${puml_file}" \
       -svg \
-    || is_success=false
+      || is_success=false
 
-  if [[ "${is_success}" = "false" ]]; then
-    failed_count=$(( failed_count + 1 ))
-    # When it errors it seems to still create an svg so delete it if there
-    rm -f "${generated_svg_file}"
-  else 
-    # Now rename the file so we can distinguish puml generated svgs from
-    # other svgs in the gitignore
-    mv "${generated_svg_file}" "${renamed_svg_file}"
+    if [[ "${is_success}" = "false" ]]; then
+      failed_count=$(( failed_count + 1 ))
+      # When it errors it seems to still create an svg so delete it if there
+      rm -f "${generated_svg_file}"
+    else 
+      # Now rename the file so we can distinguish puml generated svgs from
+      # other svgs in the gitignore
+      mv "${generated_svg_file}" "${renamed_svg_file}"
+
+      # Checksum the file so we can check next time we run
+      echo -e "${GREEN}Creating SHA1 checksum file ${BLUE}${sha1_filename}${NC}"
+      sha1sum "${puml_file}" > "${sha1_file}"
+    fi
+  else
+    echo -e "${GREEN}Skipping conversion${NC}"
   fi
 }
 
@@ -116,6 +148,7 @@ main() {
   local failed_count=0
 
   for arg in "$@"; do
+    echo
     echo -e "${GREEN}Processing argument ${BLUE}${arg}${NC}"
     # remove any trailing slash
     local file_or_dir="${arg}"; shift
@@ -150,8 +183,6 @@ main() {
       fi
     fi
   done
-  
-
 
   if [[ "${failed_count}" -gt 0 ]]; then
     echo -e "${RED}ERROR${NC}: Failed to convert ${failed_count} files" >&2
